Move WeightInfo, Config, Event into NFT pallet.
