+++ pallets/settlement/src/lib.rs
@@ -70,7 +70,7 @@ use sp_std::vec;
 use pallet_asset::MandatoryMediators;
 use pallet_base::{ensure_string_limited, try_next_post};
 pub use polymesh_common_utilities::traits::settlement::{Event, RawEvent, WeightInfo};
-use polymesh_common_utilities::traits::{asset, compliance_manager, nft, CommonConfig};
+use polymesh_common_utilities::traits::{asset, compliance_manager, CommonConfig};
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::constants::queue_priority::SETTLEMENT_INSTRUCTION_EXECUTION_PRIORITY;
 use polymesh_primitives::settlement::{
@@ -106,7 +106,7 @@ pub trait Config:
     + compliance_manager::Config
     + frame_system::Config
     + pallet_identity::Config
-    + nft::Config
+    + pallet_nft::Config
     + pallet_timestamp::Config
 {
     /// The overarching event type.
